python/requirements.txt
aspy.yaml==1.3.0
boto3==1.13.7
certifi==2019.9.11
cfgv==2.0.1
chardet==3.0.4
Click==7.0
cycler==0.10.0
httplib2==0.14.0
identify==1.4.7
idna==2.8
importlib-metadata==0.23
kiwisolver==1.1.0
matplotlib==3.1.2
more-itertools==7.2.0
nodeenv==1.3.3
numpy==1.17.4
oauth2client==4.1.3
pandas==0.25.3
pre-commit==1.20.0
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.7
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
rsa==4.0
scipy==1.4.1
seaborn==0.10.0
Shapely==1.6.4.post2
six==1.13.0
toml==0.10.0
urllib3==1.25.7
virtualenv==16.7.7
xdg==4.0.1
xlrd==1.2.0
zipp==0.6.0
